CARS/requirements.txt
numpy == 1.18.5
Flask == 1.1.2
gunicorn == 20.0.4
tensorflow-cpu == 2.3.0
Pillow == 8.1.0
streamlit == 0.74.0